moving resposne and requests types into their own file
