+++ src/services/sam/supports/types/ticketReponseTypes.ts
@@ -1,16 +1,3 @@
-export interface CreateTicketRequest {
-  ticket: {
-    subject: string,
-    priority: string,
-  }
-}
-export interface CreateContactRequest {
-  email: string,
-  first_name: string,
-  last_name: string,
-  phone_number: string
-}
-
 export interface CreateTicketResponse {
   ticket_id: string,
   subject: string,
