+++ src/services/sam/supports/types/ticketRequestTypes.ts
@@ -1,0 +1,12 @@
+export interface CreateTicketRequest {
+  ticket: {
+    subject: string,
+    priority: string,
+  }
+}
+export interface CreateContactRequest {
+  email: string,
+  first_name: string,
+  last_name: string,
+  phone_number: string
+}
